+++ bobtemplates/plone/viewlet.py
@@ -88,7 +88,7 @@ def _update_viewlets_configure_zcml(configurator):
      name="{0}"
      for="{1}"
      manager="plone.app.layout.viewlets.interfaces.IAboveContentTitle"
-     layer="{2}.interfaces.{3}"
+     layer="{2}.interfaces.I{3}"
      class=".{4}.{5}"
      template="{6}.pt"
      permission="zope2.View"
@@ -97,7 +97,7 @@ def _update_viewlets_configure_zcml(configurator):
             configurator.variables["viewlet_name"],
             iface_name,
             configurator.variables["package.dottedname"],
-            configurator.variables["browser_layer"],
+            configurator.variables["package.browserlayer"],
             configurator.variables["viewlet_python_file_name"],
             configurator.variables["viewlet_python_class_name"],
             configurator.variables["viewlet_template_name"],
@@ -109,7 +109,7 @@ def _update_viewlets_configure_zcml(configurator):
      name="{0}"
      for="{1}"
      manager="plone.app.layout.viewlets.interfaces.IAboveContentTitle"
-     layer="{2}.interfaces.{3}"
+     layer="{2}.interfaces.I{3}"
      class=".{4}.{5}"
      permission="zope2.View"
      />
@@ -117,7 +117,7 @@ def _update_viewlets_configure_zcml(configurator):
             configurator.variables["viewlet_name"],
             iface_name,
             configurator.variables["package.dottedname"],
-            configurator.variables["browser_layer"],
+            configurator.variables["package.browserlayer"],
             configurator.variables["viewlet_python_file_name"],
             configurator.variables["viewlet_python_class_name"],
         )
@@ -163,11 +163,6 @@ def prepare_renderer(configurator):
         "viewlet_python_file_name"
     ] = viewlet_python_file_name  # NOQA: E501
     configurator.target_directory = configurator.variables["package_folder"]
-    package_name = configurator.variables["package.dottedname"].replace(
-        ".", "_"
-    )  # NOQA: E501
-    browser_layer = cc.pascalcase(package_name)
-    configurator.variables["browser_layer"] = "I{0}Layer".format(browser_layer)
 
 
 def post_renderer(configurator):